requirements.txt
Django==4.0.5
djangorestframework==3.13.1
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1